# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Zarumi Anime Streaming API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type WorkType = typeof WorkType[keyof typeof WorkType];


export const WorkType = {
  tv: 'tv',
  movie: 'movie',
} as const;

export type WorkStatus = typeof WorkStatus[keyof typeof WorkStatus];


export const WorkStatus = {
  ongoing: 'ongoing',
  completed: 'completed',
  upcoming: 'upcoming',
} as const;

export interface Work {
  id: number;
  tmdbId: number;
  title: string;
  /** @nullable */
  originalTitle?: string | null;
  /** @nullable */
  synopsis?: string | null;
  /** @nullable */
  posterPath?: string | null;
  /** @nullable */
  backdropPath?: string | null;
  /** @nullable */
  customBannerUrl?: string | null;
  /** @nullable */
  customThumbnailUrl?: string | null;
  type: WorkType;
  status: WorkStatus;
  genres?: string[];
  /** @nullable */
  releaseYear?: number | null;
  /** @nullable */
  totalEpisodes?: number | null;
  /** @nullable */
  rating?: number | null;
  isFeatured?: boolean;
  viewCount?: number;
  createdAt?: string;
}

export type WorkInputType = typeof WorkInputType[keyof typeof WorkInputType];


export const WorkInputType = {
  tv: 'tv',
  movie: 'movie',
} as const;

export type WorkInputStatus = typeof WorkInputStatus[keyof typeof WorkInputStatus];


export const WorkInputStatus = {
  ongoing: 'ongoing',
  completed: 'completed',
  upcoming: 'upcoming',
} as const;

export interface WorkInput {
  tmdbId: number;
  title: string;
  /** @nullable */
  originalTitle?: string | null;
  /** @nullable */
  synopsis?: string | null;
  /** @nullable */
  posterPath?: string | null;
  /** @nullable */
  backdropPath?: string | null;
  /** @nullable */
  customBannerUrl?: string | null;
  /** @nullable */
  customThumbnailUrl?: string | null;
  type: WorkInputType;
  status?: WorkInputStatus;
  genres?: string[];
  /** @nullable */
  releaseYear?: number | null;
  /** @nullable */
  totalEpisodes?: number | null;
  /** @nullable */
  rating?: number | null;
  isFeatured?: boolean;
}

export interface WorkUpdate {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  synopsis?: string | null;
  /** @nullable */
  customBannerUrl?: string | null;
  /** @nullable */
  customThumbnailUrl?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  isFeatured?: boolean | null;
  /** @nullable */
  posterPath?: string | null;
  /** @nullable */
  backdropPath?: string | null;
  genres?: string[];
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  viewCount?: number | null;
}

export interface WorksListResponse {
  works: Work[];
  total: number;
  page: number;
  limit: number;
}

export interface WorkWithViews {
  id: number;
  tmdbId: number;
  title: string;
  /** @nullable */
  posterPath?: string | null;
  /** @nullable */
  customThumbnailUrl?: string | null;
  type: string;
  viewCount: number;
  rank?: number;
}

export interface Episode {
  id: number;
  workId: number;
  episodeNumber: number;
  /** @nullable */
  seasonNumber?: number | null;
  title: string;
  /** @nullable */
  synopsis?: string | null;
  /** @nullable */
  thumbnailPath?: string | null;
  /** @nullable */
  customThumbnailUrl?: string | null;
  /** @nullable */
  videoSlug?: string | null;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  airDate?: string | null;
  /** @nullable */
  tmdbEpisodeId?: number | null;
}

export interface EpisodeUpdate {
  /** @nullable */
  episodeNumber?: number | null;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  synopsis?: string | null;
  /** @nullable */
  customThumbnailUrl?: string | null;
  /** @nullable */
  videoSlug?: string | null;
  /** @nullable */
  duration?: number | null;
}

export interface EpisodeInput {
  episodeNumber: number;
  /** @nullable */
  seasonNumber?: number | null;
  title: string;
  /** @nullable */
  synopsis?: string | null;
  /** @nullable */
  thumbnailPath?: string | null;
  /** @nullable */
  customThumbnailUrl?: string | null;
  /** @nullable */
  videoSlug?: string | null;
  /** @nullable */
  duration?: number | null;
  /** @nullable */
  airDate?: string | null;
  /** @nullable */
  tmdbEpisodeId?: number | null;
}

export interface ViewResult {
  viewCount: number;
}

export interface SiteStats {
  totalWorks: number;
  totalEpisodes: number;
  totalViews: number;
  ongoingCount?: number;
  completedCount?: number;
}

export type ListWorksParams = {
page?: number;
limit?: number;
/**
 * @nullable
 */
genre?: string | null;
};

export type ListRecentWorksParams = {
limit?: number;
};