# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Zarumi Anime Streaming API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: works
    description: Anime works management
  - name: analytics
    description: Analytics and stats
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /works:
    get:
      operationId: listWorks
      tags: [works]
      summary: List all anime works
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: genre
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of works
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorksListResponse"
    post:
      operationId: createWork
      tags: [works]
      summary: Create a new anime work (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Work"

  /works/recent:
    get:
      operationId: listRecentWorks
      tags: [works]
      summary: Recently added works
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: Recent works
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Work"

  /works/featured:
    get:
      operationId: listFeaturedWorks
      tags: [works]
      summary: Featured/banner works
      responses:
        "200":
          description: Featured works
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Work"

  /works/{id}:
    get:
      operationId: getWork
      tags: [works]
      summary: Get a work by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Work detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Work"
        "404":
          description: Not found
    put:
      operationId: updateWork
      tags: [works]
      summary: Update a work (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Work"
    delete:
      operationId: deleteWork
      tags: [works]
      summary: Delete a work (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /works/{id}/episodes:
    get:
      operationId: listEpisodes
      tags: [works]
      summary: List episodes of a work
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Episodes list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Episode"
    post:
      operationId: addEpisode
      tags: [works]
      summary: Add an episode (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EpisodeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Episode"

  /works/{workId}/episodes/{episodeId}:
    put:
      operationId: updateEpisode
      tags: [works]
      summary: Update an episode (admin)
      parameters:
        - name: workId
          in: path
          required: true
          schema:
            type: integer
        - name: episodeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EpisodeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Episode"
        "404":
          description: Not found
    delete:
      operationId: deleteEpisode
      tags: [works]
      summary: Delete an episode (admin)
      parameters:
        - name: workId
          in: path
          required: true
          schema:
            type: integer
        - name: episodeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /works/{id}/view:
    post:
      operationId: recordView
      tags: [analytics]
      summary: Record a view for a work
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: View recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ViewResult"

  /analytics/top10:
    get:
      operationId: getTop10
      tags: [analytics]
      summary: Top 10 most accessed works
      responses:
        "200":
          description: Top 10
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkWithViews"

  /analytics/stats:
    get:
      operationId: getSiteStats
      tags: [analytics]
      summary: Overall site stats
      responses:
        "200":
          description: Site statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteStats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Work:
      type: object
      required: [id, title, tmdbId, type, status]
      properties:
        id:
          type: integer
        tmdbId:
          type: integer
        title:
          type: string
        originalTitle:
          type: ["string", "null"]
        synopsis:
          type: ["string", "null"]
        posterPath:
          type: ["string", "null"]
        backdropPath:
          type: ["string", "null"]
        customBannerUrl:
          type: ["string", "null"]
        customThumbnailUrl:
          type: ["string", "null"]
        type:
          type: string
          enum: [tv, movie]
        status:
          type: string
          enum: [ongoing, completed, upcoming]
        genres:
          type: array
          items:
            type: string
        releaseYear:
          type: ["integer", "null"]
        totalEpisodes:
          type: ["integer", "null"]
        rating:
          type: ["number", "null"]
        isFeatured:
          type: boolean
        viewCount:
          type: integer
        createdAt:
          type: string

    WorkInput:
      type: object
      required: [tmdbId, title, type]
      properties:
        tmdbId:
          type: integer
        title:
          type: string
        originalTitle:
          type: ["string", "null"]
        synopsis:
          type: ["string", "null"]
        posterPath:
          type: ["string", "null"]
        backdropPath:
          type: ["string", "null"]
        customBannerUrl:
          type: ["string", "null"]
        customThumbnailUrl:
          type: ["string", "null"]
        type:
          type: string
          enum: [tv, movie]
        status:
          type: string
          enum: [ongoing, completed, upcoming]
        genres:
          type: array
          items:
            type: string
        releaseYear:
          type: ["integer", "null"]
        totalEpisodes:
          type: ["integer", "null"]
        rating:
          type: ["number", "null"]
        isFeatured:
          type: boolean

    WorkUpdate:
      type: object
      properties:
        title:
          type: ["string", "null"]
        synopsis:
          type: ["string", "null"]
        customBannerUrl:
          type: ["string", "null"]
        customThumbnailUrl:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        isFeatured:
          type: ["boolean", "null"]
        posterPath:
          type: ["string", "null"]
        backdropPath:
          type: ["string", "null"]
        genres:
          type: array
          items:
            type: string
        rating:
          type: ["number", "null"]
        viewCount:
          type: ["integer", "null"]

    WorksListResponse:
      type: object
      required: [works, total, page, limit]
      properties:
        works:
          type: array
          items:
            $ref: "#/components/schemas/Work"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    WorkWithViews:
      type: object
      required: [id, title, tmdbId, type, viewCount]
      properties:
        id:
          type: integer
        tmdbId:
          type: integer
        title:
          type: string
        posterPath:
          type: ["string", "null"]
        customThumbnailUrl:
          type: ["string", "null"]
        type:
          type: string
        viewCount:
          type: integer
        rank:
          type: integer

    Episode:
      type: object
      required: [id, workId, episodeNumber, title]
      properties:
        id:
          type: integer
        workId:
          type: integer
        episodeNumber:
          type: integer
        seasonNumber:
          type: ["integer", "null"]
        title:
          type: string
        synopsis:
          type: ["string", "null"]
        thumbnailPath:
          type: ["string", "null"]
        customThumbnailUrl:
          type: ["string", "null"]
        videoSlug:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        airDate:
          type: ["string", "null"]
        tmdbEpisodeId:
          type: ["integer", "null"]

    EpisodeUpdate:
      type: object
      properties:
        episodeNumber:
          type: ["integer", "null"]
        title:
          type: ["string", "null"]
        synopsis:
          type: ["string", "null"]
        customThumbnailUrl:
          type: ["string", "null"]
        videoSlug:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]

    EpisodeInput:
      type: object
      required: [episodeNumber, title]
      properties:
        episodeNumber:
          type: integer
        seasonNumber:
          type: ["integer", "null"]
        title:
          type: string
        synopsis:
          type: ["string", "null"]
        thumbnailPath:
          type: ["string", "null"]
        customThumbnailUrl:
          type: ["string", "null"]
        videoSlug:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        airDate:
          type: ["string", "null"]
        tmdbEpisodeId:
          type: ["integer", "null"]

    ViewResult:
      type: object
      required: [viewCount]
      properties:
        viewCount:
          type: integer

    SiteStats:
      type: object
      required: [totalWorks, totalEpisodes, totalViews]
      properties:
        totalWorks:
          type: integer
        totalEpisodes:
          type: integer
        totalViews:
          type: integer
        ongoingCount:
          type: integer
        completedCount:
          type: integer