# lib/api-zod/src/generated/types/episodeUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Zarumi Anime Streaming API
 * OpenAPI spec version: 0.1.0
 */

export interface EpisodeUpdate {
  /** @nullable */
  episodeNumber?: number | null;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  synopsis?: string | null;
  /** @nullable */
  customThumbnailUrl?: string | null;
  /** @nullable */
  videoSlug?: string | null;
  /** @nullable */
  duration?: number | null;
}