# lib/api-zod/src/generated/types/workUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Zarumi Anime Streaming API
 * OpenAPI spec version: 0.1.0
 */

export interface WorkUpdate {
  /** @nullable */
  title?: string | null;
  /** @nullable */
  synopsis?: string | null;
  /** @nullable */
  customBannerUrl?: string | null;
  /** @nullable */
  customThumbnailUrl?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  isFeatured?: boolean | null;
  /** @nullable */
  posterPath?: string | null;
  /** @nullable */
  backdropPath?: string | null;
  genres?: string[];
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  viewCount?: number | null;
}